test(select): migrate async
